+++ packages/fuels/src/cli/commands/deploy/deployContracts.ts
